Refactor to avoid repetition in fixing jax to scipy translation
