+++ optimism/inverse/test/test_Hyperelastic_gradient_checks.py
@@ -13,6 +13,7 @@ from optimism import Mechanics
 from optimism import Objective
 from optimism import Mesh
 from optimism.material import Neohookean
+from optimism.ScipyInterface import make_scipy_linear_function
 
 from .FiniteDifferenceFixture import FiniteDifferenceFixture
 
@@ -166,8 +167,8 @@ class NeoHookeanGlobalMeshAdjointSolveFixture(FiniteDifferenceFixture):
             n = self.dofManager.get_unknown_size()
             self.objective.p = p # have to update parameters to get precond to work
             self.objective.update_precond(Uu) # update preconditioner for use in cg (will converge in 1 iteration as long as the preconditioner is not approximate)
-            dRdu = linalg.LinearOperator((n, n), lambda V: onp.asarray(self.objective.hessian_vec(Uu, V)))
-            dRdu_decomp = linalg.LinearOperator((n, n), lambda V: onp.asarray(self.objective.apply_precond(V)))
+            dRdu = linalg.LinearOperator((n, n), make_scipy_linear_function(lambda V: self.objective.hessian_vec(Uu, V)))
+            dRdu_decomp = linalg.LinearOperator((n, n),  make_scipy_linear_function(self.objective.apply_precond))
             adjointVector = linalg.cg(dRdu, onp.array(adjointLoad, copy=False), rtol=1e-10, atol=0.0, M=dRdu_decomp)[0]
 
             gradient += df_dx
